# src/Romarr.Api.V3/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Romarr",
    "description": "Romarr API docs - The v3 API docs apply to both v3 and v4 versions of Romarr. Some functionality may only be available in v4 of the Romarr application.",
    "license": {
      "name": "GPL-3.0",
      "url": "https://github.com/Romarr/Romarr/blob/develop/LICENSE"
    },
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "{protocol}://{hostpath}",
      "variables": {
        "protocol": {
          "default": "http",
          "enum": [
            "http",
            "https"
          ]
        },
        "hostpath": {
          "default": "localhost:9797"
        }
      }
    }
  ],
  "paths": {
...
    "/api/v3/importlist/bulk": {
      "put": {
        "tags": [
          "ImportList"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportListBulkResource"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportListResource"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ImportList"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportListBulkResource"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v3/importlistexclusion/paged": {
      "get": {
        "tags": [
          "ImportListExclusion"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortKey",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportListExclusionResourcePagingResource"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v3/importlistexclusion/bulk": {
      "delete": {
        "tags": [
          "ImportListExclusion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportListExclusionBulkResource"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportListExclusionBulkResource"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ImportListExclusionBulkResource"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplyTags": {
        "enum": [
          "add",
          "remove",
          "replace"
        ],
        "type": "string"
      },
...
      "Field": {
        "type": "object",
        "properties": {
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "helpText": {
            "type": "string",
            "nullable": true
          },
          "helpTextWarning": {
            "type": "string",
            "nullable": true
          },
          "helpLink": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "advanced": {
            "type": "boolean"
          },
          "selectOptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SelectOption"
            },
            "nullable": true
          },
          "selectOptionsProviderAction": {
            "type": "string",
            "nullable": true
          },
          "section": {
            "type": "string",
            "nullable": true
          },
          "hidden": {
            "type": "string",
            "nullable": true
          },
          "privacy": {
            "$ref": "#/components/schemas/PrivacyLevel"
          },
          "placeholder": {
            "type": "string",
            "nullable": true
          },
          "isFloat": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ImportListBulkResource": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "applyTags": {
            "$ref": "#/components/schemas/ApplyTags"
          },
          "enableAutomaticAdd": {
            "type": "boolean",
            "nullable": true
          },
          "rootFolderPath": {
            "type": "string",
            "nullable": true
          },
          "qualityProfileId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ImportListExclusionBulkResource": {
        "type": "object",
        "properties": {
          "ids": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ImportListExclusionResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "igdbId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ImportListExclusionResourcePagingResource": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "sortKey": {
            "type": "string",
            "nullable": true
          },
          "sortDirection": {
            "$ref": "#/components/schemas/SortDirection"
          },
          "totalRecords": {
            "type": "integer",
            "format": "int32"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportListExclusionResource"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ImportListResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Field"
            },
            "nullable": true
          },
          "implementationName": {
            "type": "string",
            "nullable": true
          },
          "implementation": {
            "type": "string",
            "nullable": true
          },
          "configContract": {
            "type": "string",
            "nullable": true
          },
          "infoLink": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "$ref": "#/components/schemas/ProviderMessage"
          },
          "tags": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "presets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportListResource"
            },
            "nullable": true
          },
          "enableAutomaticAdd": {
            "type": "boolean"
          },
          "searchForMissingEpisodes": {
            "type": "boolean"
          },
          "shouldMonitor": {
            "$ref": "#/components/schemas/MonitorTypes"
          },
          "monitorNewItems": {
            "$ref": "#/components/schemas/NewItemMonitorTypes"
          },
          "rootFolderPath": {
            "type": "string",
            "nullable": true
          },
          "qualityProfileId": {
            "type": "integer",
            "format": "int32"
          },
          "seriesType": {
            "$ref": "#/components/schemas/GameTypes"
          },
          "platformFolder": {
            "type": "boolean"
          },
          "listType": {
            "$ref": "#/components/schemas/ImportListType"
          },
          "listOrder": {
            "type": "integer",
            "format": "int32"
          },
          "minRefreshInterval": {
            "type": "string",
            "format": "date-span"
          }
        },
        "additionalProperties": false
      },
      "ImportListType": {
        "enum": [
          "program",
          "plex",
          "trakt",
          "simkl",
          "other",
          "advanced"
        ],
        "type": "string"
      },
...
      "MonitorTypes": {
        "enum": [
          "unknown",
          "all",
          "future",
          "missing",
          "existing",
          "firstSeason",
          "lastSeason",
          "latestSeason",
          "pilot",
          "recent",
          "monitorSpecials",
          "unmonitorSpecials",
          "none",
          "skip"
        ],
        "type": "string"
      },
...
      "NewItemMonitorTypes": {
        "enum": [
          "all",
          "none"
        ],
        "type": "string"
      },
...
      "PrivacyLevel": {
        "enum": [
          "normal",
          "password",
          "apiKey",
          "userName"
        ],
        "type": "string"
      },
...
      "ProviderMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/ProviderMessageType"
          }
        },
        "additionalProperties": false
      },
      "ProviderMessageType": {
        "enum": [
          "info",
          "warning",
          "error"
        ],
        "type": "string"
      },
...
      "SelectOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "hint": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GameTypes": {
        "enum": [
          "standard",
          "daily",
          "anime"
        ],
        "type": "string"
      },
      "SortDirection": {
        "enum": [
          "default",
          "ascending",
          "descending"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "X-Api-Key": {
        "type": "apiKey",
        "description": "Apikey passed as header",
        "name": "X-Api-Key",
        "in": "header"
      },
      "apikey": {
        "type": "apiKey",
        "description": "Apikey passed as query parameter",
        "name": "apikey",
        "in": "query"
      }
    }
  },
  "security": [
    {
      "X-Api-Key": [ ]
    },
    {
      "apikey": [ ]
    }
  ]
}